rename: 의상 속성 패키지 이름 변경
